Incomplete Work

+++ Lessons/Lesson3/BrokenCalculator.cs
@@ -13,19 +13,19 @@ namespace Lesson3
             }
         }
 
-        public void Add(double num1)
+        public void Add(int num1)
         {
             currentValue = currentValue + num1;
         }
-        public void Subtract(double num1)
+        public void Subtract(int num1)
         {
             currentValue = currentValue - num1;
         }
-        public void Multiplication(double num1)
+        public void Multiplication(int num1)
         {
             currentValue = currentValue * num1;
         }
-        public void Division(double num1)
+        public void Division(int num1)
         {
             currentValue = currentValue / num1;
         }
@@ -33,6 +33,3 @@ namespace Lesson3
     }
 
 }
-
-//change the calculator so you enter each value individually like a new calculator
-//add a private field to use the break the calculator
